Starting over once again, this time with Kivy.
See, what happens here is, I want to learn new technologies, so I start writing a new task manager, because nobody has ever written a decent one in my opinion. But the todo list never gets finished, probably because the technologies being tested totally suck or I run out of time or I test somebody else's product and I'm ok with it for a while...

So let's see how far we get with Kivy! The purpose is to code, not to complete.